change argument type for int in coordinates class

+++ backend/src/main/java/com/joffrey/bolber/business/domain/driver/Coordinates.java
@@ -1,6 +1,6 @@
 package com.joffrey.bolber.business.domain.driver;
 
-public record Coordinates(float horizontal, float vertical) {
+public record Coordinates(int horizontal, int vertical) {
 
 
 }
